backend/requirements.txt
Google Cloud Services
google-cloud-speech

HTTP Requests
requests

Server
Flask
Flask-Cors

Other utilities
python-dotenv